// src/enums/onboarding.enums.ts
export enum AmazonOnboardingEnum {
  TITLE = 'Accelerate Your Growth on Amazon',
  SUBTITLE = 'Connect your accounts for maximum reach & impact today',
  DESCRIPTION_1 = `Connect your ad account for complete visibility into your ad metrics,
            allowing for better performance tracking & campaign optimization.`,
  DESCRIPTION_2 = `Connect your seller central account to seamlessly sync your product listings,
            track inventory, profit margins, sales and more.`,
  DESCRIPTION_3 = `Connect your vendor central account to seamlessly sync your product listings,
            track inventory, profit margins, sales and more.`,
}

export enum WalmartOnboardingEnum {
  TITLE = 'Accelerate Your Growth on Walmart',
  SUBTITLE = 'Connect your accounts for maximum reach and impact today',
  WALMART_ADS_DESCRIPTION = `Connect your product catalog to seamlessly sync your product listings,
            track inventory, profit margins, sales and more.`,
  WALMART_CONNECT_DESCRIPTION = `Connect your ad account for complete visibility into your ad metrics, allowing for better performance tracking and campaign optimization.`,
  WALMART_SUPPLIER_DESCRIPTION = `Connect your supplier product catalog to seamlessly sync your product listings, track inventory, sales and more.`,
}

export enum AmazonPartnerTypeEnum {
  SELLER = 'SELLER',
  ADVERTISING = 'ADVERTISING',
}

export enum AccountRegionEnum {
  NORTH_AMERICA = 'North America',
  EUROPE = 'Europe',
  FAR_EAST = 'Far East',
}
export enum OnboardingTaskStatusEnum {
  PENDING = 'pending',
  COMPLETED = 'completed',
}
export enum OnboardingTypeEnum {
  AMAZON_SP_VENDOR = 'amazonSPVendor',
  AMAZON_SP_SELLER = 'amazonSPSeller',
  AMAZON_ADS = 'amazonAds',
  WALMART_ADS = 'walmartAds',
  WALMART_CONNECT = 'walmartConnect',
  WALMART_SUPPLIER = 'walmartSupplier',
}